requirements.txt
alpaca-py==0.13.6
annotated-types==0.6.0
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
frozendict==2.4.0
html5lib==1.1
idna==3.6
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.2
msgpack==1.0.7
multitasking==0.0.11
numpy==1.26.3
packaging==23.2
pandas==2.1.4
peewee==3.17.0
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sseclient-py==1.8.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
websockets==11.0.3
yfinance==0.2.35
